+++ example/lib/main.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: avoid_redundant_argument_values
+
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:pull_down_button/pull_down_button.dart';
@@ -28,25 +30,19 @@ class _MyAppState extends State<MyApp> {
 
   @override
   Widget build(BuildContext context) => MaterialApp(
-        title: 'Pull-Down button demo',
+        title: 'PullDownButton Example',
         theme: ThemeData(
           cupertinoOverrideTheme: const CupertinoThemeData(
-            scaffoldBackgroundColor: CupertinoColors.systemBackground,
-            barBackgroundColor: CupertinoDynamicColor.withBrightness(
-              color: Color(0xF0F9F9F9),
-              darkColor: Color(0xF01D1D1D),
-            ),
+            scaffoldBackgroundColor: CupertinoColors.systemGroupedBackground,
+            barBackgroundColor: Color(0xF0F9F9F9),
           ),
         ),
         darkTheme: ThemeData(
           brightness: Brightness.dark,
           cupertinoOverrideTheme: const CupertinoThemeData(
             brightness: Brightness.dark,
-            scaffoldBackgroundColor: CupertinoColors.systemBackground,
-            barBackgroundColor: CupertinoDynamicColor.withBrightness(
-              color: Color(0xF0F9F9F9),
-              darkColor: Color(0xF01D1D1D),
-            ),
+            scaffoldBackgroundColor: CupertinoColors.systemGroupedBackground,
+            barBackgroundColor: Color(0xF01D1D1D),
           ),
         ),
         themeMode: mode,
@@ -60,68 +56,221 @@ class MyHomePage extends StatelessWidget {
   final GestureTapCallback onThemeModeChange;
 
   @override
-  Widget build(BuildContext context) => CupertinoPageScaffold(
-        navigationBar: CupertinoNavigationBar(
-          leading: CupertinoButton(
-            onPressed: onThemeModeChange,
-            padding: EdgeInsets.zero,
-            child: const Icon(CupertinoIcons.sun_max_fill),
-          ),
-          middle: Text(
-            'Pull-down button example',
-            style: TextStyle(
-              color: CupertinoDynamicColor.resolve(
-                CupertinoColors.label,
-                context,
-              ),
+  Widget build(BuildContext context) {
+    final padding = EdgeInsets.only(
+      left: 16 + MediaQuery.of(context).padding.left,
+      top: MediaQuery.of(context).padding.top +
+          kMinInteractiveDimensionCupertino +
+          24,
+      right: 16 + MediaQuery.of(context).padding.right,
+      bottom: 24,
+    );
+
+    return CupertinoPageScaffold(
+      navigationBar: CupertinoNavigationBar(
+        leading: CupertinoButton(
+          onPressed: onThemeModeChange,
+          padding: EdgeInsets.zero,
+          child: const Icon(CupertinoIcons.sun_max_fill),
+        ),
+        middle: Text(
+          'PullDownButton Example',
+          style: TextStyle(
+            color: CupertinoDynamicColor.resolve(
+              CupertinoColors.label,
+              context,
             ),
           ),
         ),
-        child: const Align(
-          alignment: Alignment.centerRight,
-          child: PullDownButtonPositionUnder(),
+        trailing: _Example(
+          position: PullDownMenuPosition.under,
+          builder: (_, showMenu) => CupertinoButton(
+            onPressed: showMenu,
+            padding: EdgeInsets.zero,
+            child: const Icon(CupertinoIcons.ellipsis_circle),
+          ),
         ),
-      );
+      ),
+      child: Column(
+        children: [
+          Expanded(
+            child: ListView.separated(
+              padding: padding,
+              reverse: true,
+              itemBuilder: (context, index) {
+                final isSender = index.isEven;
+
+                return Align(
+                  alignment:
+                      isSender ? Alignment.centerRight : Alignment.centerLeft,
+                  child: _Example(
+                    position: PullDownMenuPosition.automatic,
+                    applyOpacity: false,
+                    builder: (_, showMenu) => CupertinoButton(
+                      onPressed: showMenu,
+                      padding: EdgeInsets.zero,
+                      child: _MessageExample(isSender: isSender),
+                    ),
+                  ),
+                );
+              },
+              separatorBuilder: (_, __) => const SizedBox(height: 16),
+              itemCount: 20,
+            ),
+          ),
+          ColoredBox(
+            color: CupertinoColors.systemGrey5.resolveFrom(context),
+            child: SafeArea(
+              top: false,
+              child: Padding(
+                padding: const EdgeInsets.symmetric(horizontal: 8),
+                child: Row(
+                  children: [
+                    Expanded(
+                      child: _Example(
+                        position: PullDownMenuPosition.above,
+                        builder: (_, showMenu) => CupertinoButton(
+                          onPressed: showMenu,
+                          child: const Text('Show menu above'),
+                        ),
+                      ),
+                    ),
+                    Expanded(
+                      child: _Example(
+                        position: PullDownMenuPosition.over,
+                        builder: (_, showMenu) => CupertinoButton(
+                          onPressed: showMenu,
+                          child: const Text('Show menu over'),
+                        ),
+                      ),
+                    )
+                  ],
+                ),
+              ),
+            ),
+          )
+        ],
+      ),
+    );
+  }
 }
 
+// iOS Files app menu replica
 @immutable
-class PullDownButtonPositionUnder extends StatelessWidget {
-  const PullDownButtonPositionUnder({super.key});
+class _Example extends StatelessWidget {
+  const _Example({
+    required this.position,
+    required this.builder,
+    this.applyOpacity = true,
+  });
 
-  void action() {}
-  void action2() {}
-  void deleteAction() {}
+  final PullDownMenuPosition position;
+  final PullDownMenuButtonBuilder builder;
+  final bool applyOpacity;
 
   @override
   Widget build(BuildContext context) => PullDownButton(
         itemBuilder: (context) => [
-          const PullDownMenuTitle(title: Text('Pull-Down menu')),
-          const PullDownMenuDivider(),
-          SelectablePullDownMenuItem(
-            title: 'Order by size',
-            selected: true,
-            onTap: () => action(),
-            icon: CupertinoIcons.chevron_down,
+          PullDownMenuItem(
+            enabled: false,
+            title: 'Select',
+            onTap: () {},
+            icon: CupertinoIcons.checkmark_circle,
           ),
           const PullDownMenuDivider(),
-          SelectablePullDownMenuItem(
-            title: 'Order by weight',
-            selected: false,
-            onTap: () => action2(),
+          PullDownMenuItem(
+            title: 'Connect to remote server',
+            onTap: () {},
+            icon: CupertinoIcons.cloud_upload,
           ),
           const PullDownMenuDivider.large(),
-          PullDownMenuItem(
-            title: 'Delete',
-            icon: CupertinoIcons.delete,
-            onTap: () => deleteAction(),
-            isDestructive: true,
+          SelectablePullDownMenuItem(
+            title: 'Grid',
+            selected: true,
+            onTap: () {},
+            icon: CupertinoIcons.square_grid_2x2,
           ),
+          const PullDownMenuDivider(),
+          SelectablePullDownMenuItem(
+            title: 'List',
+            selected: false,
+            onTap: () {},
+            icon: CupertinoIcons.list_bullet,
+          ),
+          const PullDownMenuDivider.large(),
+          ...PullDownMenuDivider.wrapWithDivider([
+            SelectablePullDownMenuItem(
+              title: 'Name',
+              selected: false,
+              onTap: () {},
+            ),
+            SelectablePullDownMenuItem(
+              title: 'Type',
+              selected: false,
+              onTap: () {},
+            ),
+            SelectablePullDownMenuItem(
+              title: 'Date',
+              selected: true,
+              icon: CupertinoIcons.chevron_down,
+              onTap: () {},
+            ),
+            SelectablePullDownMenuItem(
+              title: 'Size',
+              selected: false,
+              onTap: () {},
+            ),
+            SelectablePullDownMenuItem(
+              title: 'Tags',
+              selected: false,
+              onTap: () {},
+            ),
+          ]),
         ],
-        position: PullDownMenuPosition.under,
-        buttonBuilder: (context, showMenu) => CupertinoButton(
-          onPressed: showMenu,
-          padding: EdgeInsets.zero,
-          child: const Icon(CupertinoIcons.ellipsis_circle),
+        applyOpacity: applyOpacity,
+        position: position,
+        buttonBuilder: builder,
+      );
+}
+
+// Eyeballed message box from iMessage
+@immutable
+class _MessageExample extends StatelessWidget {
+  const _MessageExample({
+    required this.isSender,
+  });
+
+  final bool isSender;
+
+  @override
+  Widget build(BuildContext context) => SizedBox(
+        width: 267,
+        child: Material(
+          color: isSender
+              ? CupertinoColors.systemBlue.resolveFrom(context)
+              : CupertinoColors.systemFill.resolveFrom(context),
+          shape: const RoundedRectangleBorder(
+            borderRadius: BorderRadius.all(
+              Radius.circular(18),
+            ),
+          ),
+          child: Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 6),
+            child: Text(
+              isSender
+                  ? 'How’s next Tuesday? Can’t wait to see you! 🤗'
+                  : 'Let’s get lunch! When works for you? 😋',
+              style: TextStyle(
+                fontSize: 17,
+                height: 22 / 17,
+                letterSpacing: -0.41,
+                fontFamily: '.SF Pro Text',
+                color: isSender
+                    ? CupertinoColors.label.darkColor
+                    : CupertinoColors.label.resolveFrom(context),
+              ),
+            ),
+          ),
         ),
       );
 }