+++ lib/src/items/title.dart
@@ -5,9 +5,9 @@ import '../theme/theme.dart';
 import 'entry.dart';
 
 /// The (optional) title of the pull-down menu that is usually displayed at the
-/// top of Pull-Down menu.
+/// top of pull-down menu.
 ///
-/// Typically a [Text] widget.
+/// [title] is typically a [Text] widget.
 @immutable
 class PullDownMenuTitle extends PullDownMenuEntry {
   /// Creates a title for a pull-down menu.